Struct roach::Transaction

pub struct Transaction<'a, M: Mutability, D> { /* private fields */ }

Facilitates an atomic interaction with an Archive.

Implementations

impl<'a, M: Mutability, D> Transaction<'a, M, D>

pub fn get<K: 'static + ?Sized>( &self, key: &K ) -> Result<Option<<<D as ArchiveType<K>>::Value as DataLoad<'_>>::OutputType>, ArchiveError>where D: ArchiveType<K>,

Gets the value with the specified key from the archive, if any.

pub fn keys<K: 'static + ?Sized>( &self ) -> impl '_ + Iterator<Item = Result<<<D as ArchiveType<K>>::Key as BinaryConverter<'_>>::ValueOutput, ArchiveError>>where D: ArchiveType<K>,

Obtains an iterator over all keys of the given type in the archive.

pub fn values<K: 'static + ?Sized>( &self ) -> impl '_ + Iterator<Item = Result<<<D as ArchiveType<K>>::Value as DataLoad<'_>>::OutputType, ArchiveError>>where D: ArchiveType<K>,

Obtains an iterator over all values for the provided key type in the archive.

pub fn iter<K: 'static + ?Sized>( &self ) -> impl '_ + Iterator<Item = Result<(<<D as ArchiveType<K>>::Key as BinaryConverter<'_>>::ValueOutput, <<D as ArchiveType<K>>::Value as DataLoad<'_>>::OutputType), ArchiveError>>where D: ArchiveType<K>,

Obtains an iterator over all keys and values of the given type in the archive.

impl<'a, D> Transaction<'a, Mut, D>

pub fn commit(self) -> Result<(), ArchiveError>

Commits this transaction to the archive. Any modifications will become visible for subsequent reads and writes.

pub fn set<'b, K: 'static + ?Sized, V: 'b + ?Sized>( &'b mut self, key: &K, value: &V ) -> Result<(), ArchiveError>where D: ArchiveType<K>, <D as ArchiveType<K>>::Value: BinaryConverter<'b, Target = V>, <<<<D as ArchiveType<K>>::Value as BinaryConverter<'b>>::ByteOutput as IntoBytes>::RefType as RedbValue>::SelfType<'b>: Borrow<<<<D as ArchiveType<K>>::Value as DataLoad<'b>>::RawValueType as RedbValue>::SelfType<'b>>,

Sets the provided key to the given value in the archive.

pub fn replace<'b, K: 'static + ?Sized, V: 'b + ?Sized>( &'b mut self, key: &K, value: &V ) -> Result<Option<<<D as ArchiveType<K>>::Value as DataLoad<'_>>::OutputType>, ArchiveError>where D: ArchiveType<K>, <D as ArchiveType<K>>::Value: BinaryConverter<'b, Target = V>, <<<<D as ArchiveType<K>>::Value as BinaryConverter<'b>>::ByteOutput as IntoBytes>::RefType as RedbValue>::SelfType<'b>: Borrow<<<<D as ArchiveType<K>>::Value as DataLoad<'b>>::RawValueType as RedbValue>::SelfType<'b>>,

Replaces the value of the provided key, returning the old value (if any).

pub fn remove<K: 'static + ?Sized>( &mut self, key: &K ) -> Result<Option<<<D as ArchiveType<K>>::Value as DataLoad<'_>>::OutputType>, ArchiveError>where D: ArchiveType<K>,

Removes the value with the specified key from the archive.