Trait roach::DataTransform

pub trait DataTransform<'a, I: 'a + ?Sized, D: Direction> {
    type Output;

    // Required method
    fn apply(input: I) -> Result<Self::Output, ArchiveError>;
}

Implements the ability to convert archived data to in-memory data, or vice-versa.

Required Associated Types

type Output

The output type of this transformation.

Required Methods

fn apply(input: I) -> Result<Self::Output, ArchiveError>

Attempts to convert the given input type to the output.

Object Safety

This trait is not object safe.